remove standalone History page and move into TaskList
