--- plugins/self-service/src/components/History/index.ts
@@ -1,1 +1,0 @@
-export { HistoryRoutesPage } from './HistoryRoutesPage';
